Update sensor tests
This adds a mock cleve service for mocking the database API. One of the
tests is currently failing (`test_moved_run_directory`), which is to be
expected. The sensor needs some reworking so that it mainly acts on run
IDs instead of paths, and it also need to check for updates to runs in
the database that don't have a corresponding directory on the file
system.