feat: use cleve for tracking run directory states (#9)
* Use cleve for tracking directory state

* Add cleve service

One of the reasons for adding this is the ability make mocking of API
calls easier. As a bonus, all the API-related stuff resides in one
place, and it will be easier to adapt things if (when) the API changes.

* Update sensor tests

This adds a mock cleve service for mocking the database API. One of the
tests is currently failing (`test_moved_run_directory`), which is to be
expected. The sensor needs some reworking so that it mainly acts on run
IDs instead of paths, and it also need to check for updates to runs in
the database that don't have a corresponding directory on the file
system.

* Add run ID to trigger payload

* Handle moved run directories

* Fix schema for incomplete directories

* Detect analysis directories in new run directories

* Add API key to mock cleve service

* Update README

* Remove SSH stuff from config schema

Release-As: 0.1.0