Policy for retrying adding analyses to runs on error
This fixes an issue when adding a whole bunch of runs at the same time.
Then it might happen that runs haven't been fully committed to the
database by the time the analysis is being added, resulting in an error
that the run doesn't exist. This seems to get around that.

Furthermore, I found the jinja filter `use_none` that is needed to be
able to use proper `None` values in rules. Without that I instead get
the string `"None"`, which obviously breaks things.