feat: add rules and actions for handling sequencing runs (#10)
* Add rule and action for adding a run directory

* Simplify request

* Add rule and action for updating the state of a run

* Add rule for notifying about incomplete runs

* More flexible trigger emit

Also change type -> directory_type, and also make sure to include the
runparameters file for new directories and the detailed summary file for
analysis directories.

* Add rules and actions for adding/updating analyses

* Policy for retrying adding analyses to runs on error

This fixes an issue when adding a whole bunch of runs at the same time.
Then it might happen that runs haven't been fully committed to the
database by the time the analysis is being added, resulting in an error
that the run doesn't exist. This seems to get around that.

Furthermore, I found the jinja filter `use_none` that is needed to be
able to use proper `None` values in rules. Without that I instead get
the string `"None"`, which obviously breaks things.

* Fix add analysis

There was an issue with having the summary file being optional that has
now been fixed by forcing the content type to multipart/form-data. This
was achieved by putting the plain text parameters into files instead of
in payload.

* Proper output from actions

* Add run directories with state complete

* Rule documentation

* Action documentation

* Fix incomplete directory notification

The reason for it not working was that I hadn't looked close enough at
the action parameters for email.send_email, and apparently the mistakes
I made broke the JSON parsing that happens before the validation.

* Emit only one trigger per incomplete directory

* Add notification email addresses to config

* Use a better sender email for notifications

* Add notifiction email to README