fix: catch StopIteration when checking analysis dir
For some reason we sometimes see a StopIteration exception when checking
the analysis directories. The existence of the directory is checked
befor this, so I don't really understand where the error comes from.
This will now print a warning and return from the function in order to
prevent the sensor from crashing.